Updating sentry native to 0.4.9
