+++ vendors/info.txt
@@ -1,7 +1,7 @@
-sentry (0.4.7) (Qt IFW v1.12.0)
+sentry (0.4.9) (Qt IFW v1.15.0)
 ------
-Comes from release v3.0.0 of eve-launcher-vendor-libraries (https://github.com/ccpgames/eve-launcher-vendor-libraries/releases/tag/v3.0.0)
-Comes from TeamCity build tagged v3.0.0 (https://teamcity.evetech.net/buildConfiguration/Applications_EveLauncher_VendorLibraries_Packagers_SentryNativeMt?mode=builds)
+Comes from release v4.0.0 of eve-launcher-vendor-libraries (https://github.com/ccpgames/eve-launcher-vendor-libraries/releases/tag/v4.0.0)
+Comes from TeamCity build tagged v4.0.0 (https://teamcity.evetech.net/buildConfiguration/Applications_EveLauncher_VendorLibraries_Packagers_SentryNativeMt?mode=builds)
 
 pdm-proto (2.0.2) (Qt IFW v1.14.0)
 ---------
